Use `isolated_app` in the integration wf tests
Use `isolated_app` instead of `workflow_app` in the integration wf tests
is what we aim for in order to attain db isolation. However, this is NOT
working right now due to poor design of the test apps configuration.
Refactpring is needed in order to remove the technical debt we encounter
here.

Signed-off-by: Iuliana Voinea <iuliana.voinea@student.manchester.ac.uk>